Move test code to correct files
Prior to this commit, the application and calculator tests were in
the wrong source files.
