+++ samples/java-application-with-reflection/src/test/java/org/graalvm/demo/CalculatorTest.java
@@ -7,11 +7,32 @@ import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
 
-class ApplicationTest {
+class CalculatorTest {
 
     @Test
-    @DisplayName("message is hello native")
-    void usesReflection() {
-        assertEquals("Hello, native!", Application.getMessage());
+    @DisplayName("1 + 1 = 2")
+    void addsTwoNumbers() {
+        Calculator calculator = new Calculator();
+        assertEquals(2, calculator.add(1, 1), "1 + 1 should equal 2");
+    }
+
+    @Test
+    @DisplayName("1 + 2 = 3")
+    void addsTwoNumbers2() {
+        Calculator calculator = new Calculator();
+        assertEquals(3, calculator.add(1, 2), "1 + 2 should equal 3");
+    }
+
+    @ParameterizedTest(name = "{0} + {1} = {2}")
+    @CsvSource({
+            "0,    1,   1",
+            "1,    2,   3",
+            "49,  51, 100",
+            "1,  100, 101"
+    })
+    void add(int first, int second, int expectedResult) {
+        Calculator calculator = new Calculator();
+        assertEquals(expectedResult, calculator.add(first, second),
+                () -> first + " + " + second + " should equal " + expectedResult);
     }
 }
